サービス一本化、仕様に沿った実装。
