+++ mobile/src/main/kotlin/net/cosmoway/smalo/MyBootReceiver.kt
@@ -9,7 +9,6 @@ class MyBootReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action.equals(Intent.ACTION_BOOT_COMPLETED)) {
             context.startService(Intent(context, MyBeaconService::class.java))
-            context.startService(Intent(context, MyCommunicationService::class.java))
         }
     }
 }
