--- mobile/src/main/kotlin/net/cosmoway/smalo/MyCommunicationService.kt
@@ -1,284 +1,0 @@
-package net.cosmoway.smalo
-
-import android.app.Notification
-import android.app.PendingIntent
-import android.app.Service
-import android.content.Intent
-import android.content.IntentFilter
-import android.content.SharedPreferences
-import android.media.Ringtone
-import android.media.RingtoneManager
-import android.net.Uri
-import android.net.nsd.NsdManager
-import android.net.nsd.NsdServiceInfo
-import android.os.AsyncTask
-import android.os.Handler
-import android.os.IBinder
-import android.os.Message
-import android.preference.PreferenceManager
-import android.support.v4.app.NotificationManagerCompat
-import android.support.v7.app.NotificationCompat
-import android.util.Log
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import java.io.IOException
-import java.security.MessageDigest
-import java.security.NoSuchAlgorithmException
-import java.util.*
-
-// バックグラウンドで、一回だけ問合開閉処理するクラス。
-// 開閉処理が為されたら、自滅させる事。
-class MyCommunicationService : Service() {
-    private var mReceiver: MyBroadcastReceiver? = null
-    private var mHost: String? = null
-    private var mHashValue: String? = null
-    // Nsd Manager
-    private var mNsdManager: NsdManager? = null
-    // UUID設定用
-    private var mId: String? = null
-    // MM
-    private var mMajor: String? = null
-    private var mMinor: String? = null
-    private var mIntentFilter: IntentFilter? = null
-
-
-    companion object {
-        private val TAG_SERVICE = "MyCommunicationService"
-        val TAG_NSD = "NSD"
-        val SERVICE_TYPE = "_xdk-app-daemon._tcp."
-        val MY_SERVICE_NAME = "smalo"
-        //val MY_SERVICE_NAME = "smalo-dev"
-        val MY_SERVICE_UUID = "51a4a738-62b8-4b26-a929-3bbac2a5ce7c"
-        //val MY_SERVICE_UUID = "dddddddddddddddddddddddddddddddd"
-
-        val MY_APP_NAME = "ＳＭＡＬＯ"
-
-    }
-
-    private fun toEncryptedHashValue(algorithmName: String, value: String): String {
-        var md: MessageDigest? = null
-        try {
-            md = MessageDigest.getInstance(algorithmName)
-        } catch (e: NoSuchAlgorithmException) {
-            e.printStackTrace()
-        }
-
-        val sb: StringBuilder = StringBuilder()
-        md!!.update(value.toByteArray())
-        for (b in md.digest()) {
-            val hex = String.format("%02x", b)
-            sb.append(hex)
-        }
-        return sb.toString()
-    }
-
-    private fun getRequest(url: String) {
-        object : AsyncTask<Void?, Void?, String?>() {
-            override fun doInBackground(vararg params: Void?): String? {
-                var result: String
-                // リクエストオブジェクトを作って
-                val request: Request = Request.Builder().url(url).get().build()
-
-                // クライアントオブジェクトを作って
-                val client: OkHttpClient = OkHttpClient()
-
-                // リクエストして結果を受け取って
-                try {
-                    val response = client.newCall(request).execute()
-                    result = response.body().string()
-                } catch (e: IOException) {
-                    e.printStackTrace()
-                    return "Connection Error"
-                }
-                // 返す
-                Log.d(TAG_SERVICE, result)
-                return result
-            }
-
-            override fun onPostExecute(result: String?) {
-                if (result != null) {
-                    makeNotification(result)
-                    if (result == "200 OK") {
-                        Log.d("result", result)
-                        // TODO:resultをブロキャスしてサービスを自滅させる
-                        val uri: Uri = RingtoneManager
-                                .getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
-                        val ringtone: Ringtone = RingtoneManager
-                                .getRingtone(applicationContext, uri)
-                        ringtone.play()
-                        sendBroadcast()
-                        // サービスを自殺させるメソッド。
-                        stopSelf()
-                    } else if (result == "locked") {
-                        Log.d(TAG_SERVICE, "message:L")
-                        //TODO:開処理リクエスト。
-                        getRequest("http:/$mHost:10080/api/locks/unlocking/$mHashValue")
-                    } else if (result == "unlocked") {
-                        Log.d(TAG_SERVICE, "message:UL")
-                        //TODO:閉処理リクエスト。
-                        getRequest("http:/$mHost:10080/api/locks/locking/$mHashValue")
-
-                    }
-                }
-            }
-        }.execute()
-    }
-
-
-    private fun sendBroadcast() {
-        val broadcastIntent: Intent = Intent()
-        broadcastIntent.putExtra("id", mId)
-        broadcastIntent.action = "UPDATE_ACTION"
-        baseContext.sendBroadcast(broadcastIntent)
-    }
-
-    private fun makeNotification(title: String) {
-
-        val builder = NotificationCompat.Builder(applicationContext)
-        builder.setSmallIcon(R.mipmap.smalo_icon)
-
-        // ノーティフィケションを生成した時のインテントを作成する
-        val notificationIntent = Intent(this, Notification::class.java)
-        val contentIntent = PendingIntent.getActivity(this, 0,
-                notificationIntent, 0)
-
-        builder.setContentTitle(title) // 1行目
-        if (title == "locked") {
-            builder.setContentText("施錠されております。")
-        } else if (title == "unlocked") {
-            builder.setContentText("解錠されております。")
-        } else if (title == "unknown") {
-            builder.setContentText("鍵の状態が判りませんでした。")
-        } else if (title == "Connection Error") {
-            builder.setContentText("通信処理が正常に終了されませんでした。\n通信環境を御確認下さい。")
-        } else if (title.indexOf("400") != -1) {
-            builder.setContentText("予期せぬエラーが発生致しました。\n開発者に御問合せ下さい。")
-        } else if (title.indexOf("403") != -1) {
-            builder.setContentText("認証に失敗致しました。\nシステム管理者に登録を御確認下さい。")
-        }
-        builder.setContentIntent(contentIntent)
-        builder.setTicker(MY_APP_NAME) // 通知到着時に通知バーに表示(4.4まで)
-        // 5.0からは表示されない
-
-        val manager = NotificationManagerCompat.from(applicationContext)
-        manager.notify(1, builder.build())
-    }
-
-    fun ensureSystemServices() {
-        mNsdManager = getSystemService(Service.NSD_SERVICE) as NsdManager
-        /*if (nsdManager == null) {
-            return
-        }*/
-    }
-
-    private fun startDiscovery() {
-        mNsdManager?.discoverServices(SERVICE_TYPE, NsdManager.PROTOCOL_DNS_SD, MyDiscoveryListener())
-    }
-
-    private fun stopDiscovery() {
-        mNsdManager?.stopServiceDiscovery(MyDiscoveryListener())
-    }
-
-    private inner class MyDiscoveryListener : NsdManager.DiscoveryListener {
-        override fun onServiceFound(serviceInfo: NsdServiceInfo) {
-            Log.i(TAG_NSD, String.format("Service found serviceInfo=%s", serviceInfo))
-            if (serviceInfo.serviceType.equals(SERVICE_TYPE) &&
-                    serviceInfo.serviceName == MY_SERVICE_NAME) {
-                mNsdManager?.resolveService(serviceInfo, MyResolveListener())
-            }
-        }
-
-        override fun onDiscoveryStarted(serviceType: String) {
-            Log.i(TAG_NSD, String.format("Discovery started serviceType=%s", serviceType))
-        }
-
-        override fun onDiscoveryStopped(serviceType: String) {
-            Log.i(TAG_NSD, String.format("Discovery stopped serviceType=%s", serviceType))
-        }
-
-        override fun onServiceLost(serviceInfo: NsdServiceInfo) {
-            Log.i(TAG_NSD, String.format("Service lost serviceInfo=%s", serviceInfo))
-        }
-
-        override fun onStartDiscoveryFailed(serviceType: String, errorCode: Int) {
-            Log.w(TAG_NSD, String.format("Failed to start discovery serviceType=%s, errorCode=%d",
-                    serviceType, errorCode))
-        }
-
-        override fun onStopDiscoveryFailed(serviceType: String, errorCode: Int) {
-            Log.w(TAG_NSD, String.format("Failed to stop discovery serviceType=%s, errorCode=%d",
-                    serviceType, errorCode))
-        }
-    }
-
-    private inner class MyResolveListener : NsdManager.ResolveListener {
-        override fun onServiceResolved(serviceInfo: NsdServiceInfo) {
-            Log.i(TAG_NSD, String.format("Service resolved serviceInfo=%s", serviceInfo.host))
-            if (serviceInfo.serviceName == MY_SERVICE_NAME) {
-                mHost = serviceInfo.host.toString()
-                //stopDiscovery()
-            }
-        }
-
-        override fun onResolveFailed(serviceInfo: NsdServiceInfo, errorCode: Int) {
-            Log.w(TAG_NSD, String.format("Failed to resolve serviceInfo=%s, errorCode=%d",
-                    serviceInfo, errorCode))
-        }
-    }
-
-    // サービスからブロードキャストされ、値を受け取った時に動かしたい内容を書く
-    private val updateHandler = object : Handler() {
-        override fun handleMessage(msg: Message) {
-            val bundle = msg.data
-            mMajor = bundle.getString("major")
-            mMinor = bundle.getString("minor")
-            Log.d(TAG_SERVICE, "major:$mMajor, minor:$mMinor")
-            mHashValue = toEncryptedHashValue("SHA-256", "$mId|$mMajor|$mMinor")
-            getRequest("http:/$mHost:10080/api/locks/status/$mHashValue")
-        }
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        Log.d(TAG_SERVICE, "Created")
-        mReceiver = MyBroadcastReceiver()
-        mIntentFilter = IntentFilter()
-        (mIntentFilter as IntentFilter).addAction("UPDATE_ACTION")
-        registerReceiver(mReceiver, mIntentFilter)
-        (mReceiver as MyBroadcastReceiver).registerHandler(updateHandler)
-        ensureSystemServices()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        unregisterReceiver(mReceiver)
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        Log.d(TAG_SERVICE, "Command Started")
-        var id: String? = intent?.extras?.getString("id")
-        val sp: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(this)
-        // 端末固有識別番号読出
-        mId = sp.getString("SaveString", null)
-        if (mId == null) {
-            Log.d("id", "null")
-            if (id != null) {
-                mId = id
-            } else {
-                // 端末固有識別番号取得
-                mId = UUID.randomUUID().toString()
-            }
-            // 端末固有識別番号記憶
-            sp.edit().putString("SaveString", mId).apply()
-        }
-        Log.d("id", mId)
-        if (mHost == null) {
-            startDiscovery()
-        }
-        return START_STICKY;
-    }
-
-    override fun onBind(intent: Intent?): IBinder? {
-        return null
-    }
-}
